重构ClientUtils获取请求中信息生成相关标识
